[redesign] #54 엔티티 이름 수정
- DogInfo -> UserInfo
- UserInfo -> ProfileInfo
