+++ SniffMeet/SniffMeet/Source/Scene/Walk/RespondWalk/View/RespondWalkViewController.swift
@@ -73,8 +73,8 @@ final class RespondWalkViewController: BaseViewController, RespondWalkViewable {
     }
     override func configureAttributes() {
         profileView.configure(
-            name: UserInfo.example.name,
-            keywords: UserInfo.example.keywords.map { $0.rawValue }
+            name: ProfileInfo.example.name,
+            keywords: ProfileInfo.example.keywords.map { $0.rawValue }
         )
         messageLabel.text = "HomeView에서 dogInfo의 변경을 알아야 하더라구요. Presenter에서 HomePresenterOutput 프로토콜을 채택하도록 설정해줬습니다."
         submitButton.setTitle(Context.abledSubmitButtonTitle, for: .normal)
